Fix MaxPayload type in ServerConfig to match BrokerConfiguration
- Changed ServerConfig.MaxPayload from int to long
- Resolves compilation error in ConfigurationMapper.cs
- Maintains consistency between BrokerConfiguration and ServerConfig models
- Required to support large payload sizes like 100GB (107,374,182,400 bytes)

+++ src/DotGnatly.Nats/Bindings/NatsConfiguration.cs
@@ -14,7 +14,7 @@ public class ServerConfig
     public int Port { get; set; } = 4222;
 
     [JsonPropertyName("max_payload")]
-    public int MaxPayload { get; set; } = 1048576; // 1MB
+    public long MaxPayload { get; set; } = 1048576; // 1MB
 
     [JsonPropertyName("max_control_line")]
     public int MaxControlLine { get; set; } = 4096;
